discord_bot/requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
uvicorn==0.34.3
pydantic==2.11.7
typing-inspection==0.4.1
typing_extensions==4.14.0
anyio==4.9.0
h11==0.16.0
httpcore==1.0.9
sniffio==1.3.1
certifi==2025.6.15
idna==3.10
starlette==0.46.2
pydantic_core==2.33.2
dotenv
python-dotenv
watchgod
aiohttp
fastapi
httpx